requirements.txt
Django==4.2.8
djangorestframework==3.14.0
gunicorn==21.2.0  # serve production server
whitenoise==6.6.0  # static files
pyodide-py==0.24.1  # Pyodide package
python-dotenv==1.0.0  # .env file to os.environ
boltons==23.1.1  # Better builtins
